requirements.txt
Flask==3.0.3
pytest==8.3.3
pytest-mock==3.14.0
python-magic==0.4.27
PyPDF2==3.0.1
Pillow==10.4.0
pytesseract==0.3.10
python-docx==1.1.2
openpyxl==3.1.5
google-genai==1.18.0
